+++ internal/longest_line.go
@@ -4,6 +4,7 @@ import (
 	"strings"
 
 	"github.com/gookit/color"
+	"github.com/mattn/go-runewidth"
 )
 
 // ReturnLongestLine returns the longest line with a given separator
@@ -11,7 +12,7 @@ func ReturnLongestLine(text, sep string) string {
 	lines := strings.Split(text, sep)
 	var longest string
 	for _, line := range lines {
-		if len(color.ClearCode(line)) > len(color.ClearCode(longest)) {
+		if runewidth.StringWidth(color.ClearCode(line)) > runewidth.StringWidth(color.ClearCode(longest)) {
 			longest = line
 		}
 	}